changed name of sim.py to sim_cli.py
